+++ app/models/Genre.scala
@@ -1,3 +1,3 @@
-package models.genres
+package models
 
 case class Genre(id: Long, name: String, parent_genre: Option[Genre])
