+++ app/models/PublishingHouse.scala
@@ -1,3 +1,3 @@
-package models.publishingHouses
+package models
 
 case class PublishingHouse(id: Long, name: String)
