--- app/models/books/BookService.scala
@@ -1,5 +1,0 @@
-package models.books
-
-import models.services.{BasicService, PageService}
-
-trait BookService extends BasicService[Book] with PageService[Book]
