--- app/models/friends/FriendService.scala
@@ -1,6 +1,0 @@
-package models.friends
-
-import models.services.{BasicService, PageService}
-
-
-trait FriendService extends BasicService[Friend] with PageService[Friend]
