--- app/models/genres/GenreService.scala
@@ -1,5 +1,0 @@
-package models.genres
-
-import models.services.{BasicService, PageService}
-
-trait GenreService extends BasicService[Genre] with PageService[Genre]
