--- app/models/publishingHouses/PublishingHouseService.scala
@@ -1,5 +1,0 @@
-package models.publishingHouses
-
-import models.services.{BasicService, PageService}
-
-trait PublishingHouseService extends BasicService[PublishingHouse] with PageService[PublishingHouse]
