+++ app/services/PageService.scala
@@ -1,4 +1,4 @@
-package models.services
+package services
 
 import models.Page
