+++ app/services/books/BookService.scala
@@ -1,0 +1,6 @@
+package services.books
+
+import models.Book
+import services.{BasicService, PageService}
+
+trait BookService extends BasicService[Book] with PageService[Book]
