+++ app/services/borrowing/BorrowingService.scala
@@ -1,10 +1,9 @@
-package models.borrowing
+package services.borrowing
 
 import java.util.Date
 
-import models.books.Book
-import models.friends.Friend
-import models.services.PageService
+import models.{Book, Borrowing, Friend}
+import services.PageService
 
 trait BorrowingService extends PageService[Borrowing]{
   def borrow(book: Book, friend: Friend, date: Date): Option[Borrowing]
