+++ app/services/friends/FriendService.scala
@@ -1,0 +1,7 @@
+package services.friends
+
+import models.Friend
+import services.{BasicService, PageService}
+
+
+trait FriendService extends BasicService[Friend] with PageService[Friend]
