+++ app/services/genres/GenreService.scala
@@ -1,0 +1,6 @@
+package services.genres
+
+import models.Genre
+import services.{BasicService, PageService}
+
+trait GenreService extends BasicService[Genre] with PageService[Genre]
