+++ app/services/publishingHouses/PublishingHouseService.scala
@@ -1,0 +1,6 @@
+package services.publishingHouses
+
+import models.PublishingHouse
+import services.{BasicService, PageService}
+
+trait PublishingHouseService extends BasicService[PublishingHouse] with PageService[PublishingHouse]
