+++ app/views/friends/createForm.scala.html
@@ -1,4 +1,4 @@
-@(friendForm: Form[models.friends.Friend])(implicit messages: Messages)
+@(friendForm: Form[Friend])(implicit messages: Messages)
 
     @import b3.vertical.fieldConstructor
