+++ app/views/friends/details.scala.html
@@ -1,4 +1,4 @@
-@import models.friends.Friend
+@import models.Friend
 @(friend: Friend)
 @main("About friend") {
     <h2>
