+++ app/views/friends/editForm.scala.html
@@ -1,4 +1,4 @@
-@(id: Long, friendForm: Form[models.friends.Friend])(implicit messages: Messages)
+@(id: Long, friendForm: Form[Friend])(implicit messages: Messages)
 
     @import b3.vertical.fieldConstructor
