+++ app/views/friends/list.scala.html
@@ -1,4 +1,4 @@
-@(currentPage: Page[models.friends.Friend], currentOrderBy: Int, currentFilter: String)(
+@(currentPage: Page[Friend], currentOrderBy: Int, currentFilter: String)(
         implicit flash: play.api.mvc.Flash, messages: Messages)
 
     @import b3.vertical.fieldConstructor
